+++ packages/cms/cmscombine/src/constants.ts
@@ -12,12 +12,6 @@ export const ATTRIBUTES = {
       list: generateDynamicAttibuteValue('list'),
 
       /**
-       * Defines the target list where all lists will be combined into.
-       * If not defined, the first queried list will act as the target.
-       */
-      target: generateDynamicAttibuteValue('target'),
-
-      /**
        * Defines an element where to display the total items of the list.
        */
       itemsCount: generateDynamicAttibuteValue('items-count'),